requirements.txt
# Dépendances pour Assistant Gazelle V5

# API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Client HTTP pour les appels API
requests>=2.31.0
httpx>=0.24.0

# Base de données - Supabase (PostgreSQL)
psycopg2-binary>=2.9.9
supabase>=2.0.0

# Gestion des variables d'environnement (.env)
python-dotenv>=1.0.0

# Recherche vectorielle et embeddings (CRITIQUE - nécessaire pour vector_search.py)
numpy>=1.24.0
openai>=1.0.0

# Tests (optionnel - pour tests d'intégration)
pytest>=7.4.0

# Interface Web pour entraînement
flask>=3.0.0

# Rapports Google Sheets
gspread>=6.0.0
google-auth>=2.29.0
apscheduler>=3.10.4
sendgrid>=6.11.0
pytz>=2024.1

dateparser>=1.1.0  # Parser de dates naturelles (français, anglais)
